requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
pydantic>=2.6
pydantic-settings>=2.2
google-generativeai>=0.8.0
yt-dlp>=2024.7.0
jinja2>=3.1
aiofiles>=23.2

